cosas

+++ Employee.h
@@ -9,7 +9,7 @@ typedef struct
     int id;
     char nombre[128];
     int horasTrabajadas;
-    int sueldo;
+    float sueldo;
 } Employee;
 
 Employee* employee_new();
